[Reference]

​ TODO

Contents

​

[Reference]

Contents

​## Index ​

TulipaEnergyModel.create_parameters_and_sets_from_file

TulipaEnergyModel.optimise_investments

TulipaEnergyModel.save_solution_to_file

TulipaEnergyModel.create_parameters_and_sets_from_file — Method

parameters, sets = create_parameters_and_sets_from_file(input_folder)

Returns two NamedTuples with all parameters and sets read and created from the input files in the input_folder.

TulipaEnergyModel.optimise_investments — Method

optimise_investments

This is a doc for optimise_investments. It should probably be improved.

TulipaEnergyModel.save_solution_to_file — Method

save_solution_to_file(output_file, v_investment, p_unit_capacity)

Saves the solution variable vinvestment to a file "investments.csv" inside `outputfile. The format of each row isa,v,p*v, whereais the asset indexingvinvestment,vis correspondingvinvestmentvalue, andpis the correspondingpunitcapacity` value.