Reference

TulipaEnergyModel.EnergyProblem

TulipaEnergyModel.GraphAssetData

TulipaEnergyModel.GraphFlowData

TulipaEnergyModel.ModelParameters

TulipaEnergyModel.RepresentativePeriod

TulipaEnergyModel.Timeframe

TulipaEnergyModel.TulipaConstraint

TulipaEnergyModel.TulipaVariable

TulipaEnergyModel.Year

TulipaEnergyModel._check_initial_storage_level!

TulipaEnergyModel._get_graph_asset_or_flow

TulipaEnergyModel._interpolate_storage_level!

TulipaEnergyModel.add_consumer_constraints!

TulipaEnergyModel.add_expression_is_charging_terms_rep_period_constraints!

TulipaEnergyModel.add_expression_terms_over_clustered_year_constraints!

TulipaEnergyModel.add_expression_terms_rep_period_constraints!

TulipaEnergyModel.add_expression_units_on_terms_rep_period_constraints!

TulipaEnergyModel.add_flow_variables!

TulipaEnergyModel.add_group_constraints!

TulipaEnergyModel.add_investment_variables!

TulipaEnergyModel.add_ramping_constraints!

TulipaEnergyModel.add_storage_constraints!

TulipaEnergyModel.add_storage_variables!

TulipaEnergyModel.add_transport_constraints!

TulipaEnergyModel.add_unit_commitment_variables!

TulipaEnergyModel.attach_coefficient!

TulipaEnergyModel.attach_constraint!

TulipaEnergyModel.attach_expression!

TulipaEnergyModel.calculate_annualized_cost

TulipaEnergyModel.calculate_salvage_value

TulipaEnergyModel.calculate_weight_for_investment_discounts

TulipaEnergyModel.calculate_weight_for_investment_discounts

TulipaEnergyModel.compute_dual_variables!

TulipaEnergyModel.create_internal_structures

TulipaEnergyModel.create_intervals_for_years

TulipaEnergyModel.create_model

TulipaEnergyModel.create_model!

TulipaEnergyModel.default_parameters

TulipaEnergyModel.duration

TulipaEnergyModel.export_solution_to_csv_files

TulipaEnergyModel.export_solution_to_csv_files

TulipaEnergyModel.filter_graph

TulipaEnergyModel.get_graph_value_or_missing

TulipaEnergyModel.is_active

TulipaEnergyModel.profile_aggregation

TulipaEnergyModel.read_parameters_from_file

TulipaEnergyModel.run_scenario

TulipaEnergyModel.safe_comparison

TulipaEnergyModel.safe_inclusion

TulipaEnergyModel.save_solution!

TulipaEnergyModel.save_solution!

TulipaEnergyModel.solve_model

TulipaEnergyModel.solve_model!

TulipaEnergyModel.tmp_create_partition_tables

TulipaEnergyModel.tmp_example_of_flow_expression_problem

TulipaEnergyModel.EnergyProblem — Type

Structure to hold all parts of an energy problem. It is a wrapper around various other relevant structures. It hides the complexity behind the energy problem, making the usage more friendly, although more verbose.

Fields

db_connection: A DuckDB connection to the input tables in the model

graph: The Graph object that defines the geometry of the energy problem.

model: A JuMP.Model object representing the optimization model.

objective_value: The objective value of the solved problem (Float64).

variables: A TulipaVariable structure to store all the information related to the variables in the model.

constraints: A TulipaConstraint structure to store all the information related to the constraints in the model.

representative_periods: A vector of Representative Periods.

solved: A boolean indicating whether the model has been solved or not.

termination_status: The termination status of the optimization model.

timeframe: A structure with the number of periods in the representative_periods and the mapping between the periods and their representatives.

model_parameters: A ModelParameters structure to store all the parameters that are exclusive of the model.

years: A vector with the information of all the milestone years.

Constructor

EnergyProblem(connection): Constructs a new EnergyProblem object with the given connection. The constraints_partitions field is computed from the representative_periods, and the other fields are initialized with default values.

See the basic example tutorial to see how these can be used.

TulipaEnergyModel.GraphAssetData — Type

Structure to hold the asset data in the graph.

TulipaEnergyModel.GraphFlowData — Type

Structure to hold the flow data in the graph.

TulipaEnergyModel.ModelParameters — Type

ModelParameters(;key = value, ...)
ModelParameters(path; ...)
ModelParameters(connection; ...)
ModelParameters(connection, path; ...)

Structure to hold the model parameters. Some values are defined by default and some required explicit definition.

If path is passed, it is expected to be a string pointing to a TOML file with a key = value list of parameters. Explicit keyword arguments take precedence.

If connection is passed, the default discount_year is set to the minimum of all milestone years. In other words, we check for the table year_data for the column year where the column is_milestone is true. Explicit keyword arguments take precedence.

If both are passed, then path has preference. Explicit keyword arguments take precedence.

Parameters

discount_rate::Float64 = 0.0: The model discount rate.

discount_year::Int: The model discount year.

TulipaEnergyModel.RepresentativePeriod — Type

Structure to hold the data of one representative period.

TulipaEnergyModel.Timeframe — Type

Structure to hold the data of the timeframe.

TulipaEnergyModel.TulipaConstraint — Type

Structure to hold the JuMP constraints for the TulipaEnergyModel

TulipaEnergyModel.TulipaVariable — Type

Structure to hold the JuMP variables for the TulipaEnergyModel

TulipaEnergyModel.Year — Type

Structure to hold the data of the year.

TulipaEnergyModel._check_initial_storage_level! — Method

_check_initial_storage_level!(df)

Determine the starting value for the initial storage level for interpolating the storage level. If there is no initial storage level given, we will use the final storage level. Otherwise, we use the given initial storage level.

TulipaEnergyModel._get_graph_asset_or_flow — Method

_get_graph_asset_or_flow(graph, a)
_get_graph_asset_or_flow(graph, (u, v))

Returns graph[a] or graph[u, v].

TulipaEnergyModel._interpolate_storage_level! — Method

_interpolate_storage_level!(df, time_column::Symbol)

Transform the storage level dataframe from grouped timesteps or periods to incremental ones by interpolation. The starting value is the value of the previous grouped timesteps or periods or the initial value. The ending value is the value for the grouped timesteps or periods.

TulipaEnergyModel.add_consumer_constraints! — Method

addconsumerconstraints!(model, graph, dataframes, Ac, incomingflowhighestinoutresolution, outgoingflowhighestinoutresolution, )

Adds the consumer asset constraints to the model.

TulipaEnergyModel.add_expression_is_charging_terms_rep_period_constraints! — Method

add_expression_is_charging_terms_rep_period_constraints!(df_cons,
                                                   is_charging_indices,
                                                   is_charging_variables,
                                                   workspace
                                                   )

Computes the is_charging expressions per row of df_cons for the constraints that are within (intra) the representative periods.

This function is only used internally in the model.

This strategy is based on the replies in this discourse thread:

https://discourse.julialang.org/t/help-improving-the-speed-of-a-dataframes-operation/107615/23

TulipaEnergyModel.add_expression_terms_over_clustered_year_constraints! — Method

add_expression_terms_over_clustered_year_constraints!(df_inter,
                                           df_flows,
                                           df_map,
                                           graph,
                                           representative_periods,
                                           )

Computes the incoming and outgoing expressions per row of df_inter for the constraints that are between (inter) the representative periods.

This function is only used internally in the model.

TulipaEnergyModel.add_expression_terms_rep_period_constraints! — Method

add_expression_terms_rep_period_constraints!(df_cons,
                                           df_flows,
                                           workspace;
                                           use_highest_resolution = true,
                                           multiply_by_duration = true,
                                           add_min_outgoing_flow_duration = false,
                                           )

Computes the incoming and outgoing expressions per row of df_cons for the constraints that are within (intra) the representative periods.

This function is only used internally in the model.

This strategy is based on the replies in this discourse thread:

https://discourse.julialang.org/t/help-improving-the-speed-of-a-dataframes-operation/107615/23

Implementation

This expression computation uses a workspace to store all variables defined for each timestep. The idea of this algorithm is to append all variables defined at time timestep in workspace[timestep] and then aggregate then for the constraint time block.

The algorithm works like this:

Loop over each group of (asset, year, rep_period)

1.1. Loop over each variable in the group: (varidx, vartimeblockstart, vartimeblockend) 1.1.1. Loop over each timestep in vartimeblockstart:vartimeblockend 1.1.1.1. Compute the coefficient of the variable based on the repperiod resolution and the variable efficiency 1.1.1.2. Store (varidx, coefficient) in workspace[timestep] 1.2. Loop over each constraint in the group: (considx, constimeblockstart, constimeblockend) 1.2.1. Aggregate all variables in workspace[timestep] for timestep in the time block to create a list of variable indices and their coefficients [(var_idx1, coef1), ...] 1.2.2. Compute the expression using the variable container, the indices and coefficients

Notes:

On step 1.2.1, the aggregation can be either by uniqueness of not, i.e., if the variable happens in more that one workspace[timestep], should we add up the coefficients or not. This is defined by the keyword

multiply_by_duration

TulipaEnergyModel.add_expression_units_on_terms_rep_period_constraints! — Method

add_expression_units_on_terms_rep_period_constraints!(
    df_cons,
    df_units_on,
    workspace,
)

Computes the units_on expressions per row of df_cons for the constraints that are within (intra) the representative periods.

This function is only used internally in the model.

This strategy is based on the replies in this discourse thread:

https://discourse.julialang.org/t/help-improving-the-speed-of-a-dataframes-operation/107615/23

TulipaEnergyModel.add_flow_variables! — Method

add_flow_variables!(model, variables)

Adds flow variables to the optimization model based on data from the variables. The flow variables are created using the @variable macro for each row in the :flows dataframe.

TulipaEnergyModel.add_group_constraints! — Method

add_group_constraints!(connection, model, ...)

Adds group constraints for assets that share a common limits or bounds

TulipaEnergyModel.add_investment_variables! — Method

add_investment_variables!(model, graph, sets)

Adds investment, decommission, and energy-related variables to the optimization model, and sets integer constraints on selected variables based on the graph data.

TulipaEnergyModel.add_ramping_constraints! — Method

add_ramping_and_unit_commitment_constraints!(model, graph, ...)

Adds the ramping constraints for producer and conversion assets where ramping = true in assets_data

TulipaEnergyModel.add_storage_constraints! — Method

addstorageconstraints!(model, graph,...)

Adds the storage asset constraints to the model.

TulipaEnergyModel.add_storage_variables! — Method

add_storage_variables!(model, ...)

Adds storage-related variables to the optimization model, including storage levels for both intra-representative periods and inter-representative periods, as well as charging state variables. The function also optionally sets binary constraints for certain charging variables based on storage methods.

TulipaEnergyModel.add_transport_constraints! — Method

addtransportconstraints!(model, graph, dfflows, flow, Ft, flowsinvestment)

Adds the transport flow constraints to the model.

TulipaEnergyModel.add_unit_commitment_variables! — Method

add_unit_commitment_variables!(model, ...)

Adds unit commitment variables to the optimization model based on the :units_on indices. Additionally, variables are constrained to be integers based on the sets structure.

TulipaEnergyModel.attach_coefficient! — Method

attach_coefficient!(cons, name, container)

Attach a coefficient named name stored in container. This checks that the container length matches the stored indices number of rows.

TulipaEnergyModel.attach_constraint! — Method

attach_constraint!(model, cons, name, container)

Attach a constraint named name stored in container, and set model[name] = container. This checks that the container length matches the stored indices number of rows.

TulipaEnergyModel.attach_expression! — Method

attach_expression!(cons, name, container)
attach_expression!(model, cons, name, container)

Attach a expression named name stored in container, and optionally set model[name] = container. This checks that the container length matches the stored indices number of rows.

TulipaEnergyModel.calculate_annualized_cost — Method

calculate_annualized_cost(discount_rate, economic_lifetime, investment_cost, years, investable_assets)

Calculates the annualized cost for each asset, both energy assets and transport assets, in each year using provided discount rates, economic lifetimes, and investment costs.

Arguments

discount_rate::Dict: A dictionary where the key is an asset or a pair of assets (asset1, asset2) for transport assets, and the value is the discount rate.

economic_lifetime::Dict: A dictionary where the key is an asset or a pair of assets (asset1, asset2) for transport assets, and the value is the economic lifetime.

investment_cost::Dict: A dictionary where the key is a tuple (year, asset) or (year, (asset1, asset2)) for transport assets, and the value is the investment cost.

years::Array: An array of years to be considered.

investable_assets::Dict: A dictionary where the key is a year, and the value is an array of assets that are relevant for that year.

Returns

A Dict where the keys are tuples (year, asset) representing the year and the asset, and the values are the calculated annualized cost for each asset in each year.

Formula

The annualized cost for each asset in year is calculated using the formula:

annualized_cost = discount_rate[asset] / (
    (1 + discount_rate[asset]) *
    (1 - 1 / (1 + discount_rate[asset])^economic_lifetime[asset])
) * investment_cost[(year, asset)]

Example for energy assets

discount_rate = Dict("asset1" => 0.05, "asset2" => 0.07)

economic_lifetime = Dict("asset1" => 10, "asset2" => 15)

investment_cost   = Dict((2021, "asset1") => 1000, (2021, "asset2") => 1500,
                         (2022, "asset1") => 1100, (2022, "asset2") => 1600)
years = [2021, 2022]

investable_assets = Dict(2021 => ["asset1", "asset2"],
                         2022 => ["asset1"])

costs = calculate_annualized_cost(discount_rate, economic_lifetime, investment_cost, years, investable_assets)

# output

Dict{Tuple{Int64, String}, Float64} with 3 entries:
  (2021, "asset1") => 123.338
  (2021, "asset2") => 153.918
  (2022, "asset1") => 135.671

Example for transport assets

discount_rate = Dict(("asset1", "asset2") => 0.05, ("asset3", "asset4") => 0.07)

economic_lifetime = Dict(("asset1", "asset2") => 10, ("asset3", "asset4") => 15)

investment_cost   = Dict((2021, ("asset1", "asset2")) => 1000, (2021, ("asset3", "asset4")) => 1500,
                         (2022, ("asset1", "asset2")) => 1100, (2022, ("asset3", "asset4")) => 1600)
years = [2021, 2022]

investable_assets = Dict(2021 => [("asset1", "asset2"), ("asset3", "asset4")],
                         2022 => [("asset1", "asset2")])

costs = calculate_annualized_cost(discount_rate, economic_lifetime, investment_cost, years, investable_assets)

# output

Dict{Tuple{Int64, Tuple{String, String}}, Float64} with 3 entries:
  (2022, ("asset1", "asset2")) => 135.671
  (2021, ("asset3", "asset4")) => 153.918
  (2021, ("asset1", "asset2")) => 123.338

TulipaEnergyModel.calculate_salvage_value — Method

calculate_salvage_value(discount_rate,
                        economic_lifetime,
                        annualized_cost,
                        years,
                        investable_assets,
                        )

Calculates the salvage value for each asset, both energy assets and transport assets.

Arguments

discount_rate::Dict: A dictionary where the key is an asset or a pair of assets (asset1, asset2) for transport assets, and the value is the discount rate.

economic_lifetime::Dict: A dictionary where the key is an asset or a pair of assets (asset1, asset2) for transport assets, and the value is the economic lifetime.

annualized_cost::Dict: A Dict where the keys are tuples (year, asset) representing the year and the asset, and the values are the annualized cost for each asset in each year.

years::Array: An array of years to be considered.

investable_assets::Dict: A dictionary where the key is a year, and the value is an array of assets that are relevant for that year.

Returns

A Dict where the keys are tuples (year, asset) representing the year and the asset, and the values are the salvage value for each asset in each year.

Formula

The salvage value for each asset in year is calculated using the formula:

salvagevalue = annualizedcost[(year, asset)] * sum( 1 / (1 + discountrate[asset])^(yearalias - year) for yearalias in salvagevalue_set[(year, asset)] )

Example for energy assets

discount_rate = Dict("asset1" => 0.05, "asset2" => 0.07)

economic_lifetime = Dict("asset1" => 10, "asset2" => 15)

annualized_cost =
    Dict((2021, "asset1") => 123.338, (2021, "asset2") => 153.918, (2022, "asset1") => 135.671)

years = [2021, 2022]

investable_assets = Dict(2021 => ["asset1", "asset2"], 2022 => ["asset1"])

salvage_value = calculate_salvage_value(
    discount_rate,
    economic_lifetime,
    annualized_cost,
    years,
    investable_assets,
)

# output
Dict{Tuple{Int64, String}, Float64} with 3 entries:
  (2021, "asset1") => 759.2
  (2021, "asset2") => 1202.24
  (2022, "asset1") => 964.325

Example for transport assets

discount_rate = Dict(("asset1", "asset2") => 0.05, ("asset3", "asset4") => 0.07)

economic_lifetime = Dict(("asset1", "asset2") => 10, ("asset3", "asset4") => 15)

annualized_cost = Dict(
    (2022, ("asset1", "asset2")) => 135.671,
    (2021, ("asset3", "asset4")) => 153.918,
    (2021, ("asset1", "asset2")) => 123.338,
)

years = [2021, 2022]

investable_assets =
    Dict(2021 => [("asset1", "asset2"), ("asset3", "asset4")], 2022 => [("asset1", "asset2")])

salvage_value = calculate_salvage_value(
    discount_rate,
    economic_lifetime,
    annualized_cost,
    years,
    investable_assets,
)

# output

Dict{Tuple{Int64, Tuple{String, String}}, Float64} with 3 entries:
  (2022, ("asset1", "asset2")) => 964.325
  (2021, ("asset3", "asset4")) => 1202.24
  (2021, ("asset1", "asset2")) => 759.2

TulipaEnergyModel.calculate_weight_for_investment_discounts — Method

calculate_weight_for_investment_discounts(social_rate,
                                          discount_year,
                                          salvage_value,
                                          investment_cost,
                                          years,
                                          investable_assets,
                                         )

Calculates the weight for investment discounts for each asset, both energy assets and transport assets.

Arguments

social_rate::Float64: A value with the social discount rate.

discount_year::Int64: A value with the discount year for all the investments.

salvage_value::Dict: A dictionary where the key is an tuple (year, asset) or (year, (asset1, asset2)) for transport assets, and the value is the salvage value.

investment_cost::Dict: A dictionary where the key is an tuple (year, asset) or (year, (asset1, asset2)) for transport assets, and the value is the investment cost.

years::Array: An array of years to be considered.

investable_assets::Dict: A dictionary where the key is a year, and the value is an array of assets that are relevant for that year.

Returns

A Dict where the keys are tuples (year, asset) representing the year and the asset, and the values are the weights for investment discounts.

Formula

The weight for investment discounts for each asset in year is calculated using the formula:

weightforinvestmentdiscounts = 1 / (1 + socialrate)^(year - discountyear) * (1 - salvagevalue[(year, asset)] / investment_cost[(year, asset)])

Example for energy assets

social_rate = 0.02

discount_year = 2000

salvage_value = Dict(
    (2021, "asset1") => 759.1978422,
    (2021, "asset2") => 1202.2339859,
    (2022, "asset1") => 964.3285406,
)

investment_cost = Dict(
    (2021, "asset1") => 1000,
    (2021, "asset2") => 1500,
    (2022, "asset1") => 1100,
    (2022, "asset2") => 1600,
)
years = [2021, 2022]

investable_assets = Dict(2021 => ["asset1", "asset2"], 2022 => ["asset1"])

weights = calculate_weight_for_investment_discounts(
    social_rate,
    discount_year,
    salvage_value,
    investment_cost,
    years,
    investable_assets,
)

# output

Dict{Tuple{Int64, String}, Float64} with 3 entries:
  (2021, "asset1") => 0.158875
  (2021, "asset2") => 0.130973
  (2022, "asset1") => 0.0797796

Example for transport assets

social_rate = 0.02

discount_year = 2000

salvage_value = Dict(
  (2022, ("asset1", "asset2")) => 964.325,
  (2021, ("asset3", "asset4")) => 1202.24,
  (2021, ("asset1", "asset2")) => 759.2,
)

investment_cost   = Dict((2021, ("asset1", "asset2")) => 1000, (2021, ("asset3", "asset4")) => 1500,
                         (2022, ("asset1", "asset2")) => 1100, (2022, ("asset3", "asset4")) => 1600)
years = [2021, 2022]

investable_assets = Dict(2021 => [("asset1", "asset2"), ("asset3", "asset4")],
                         2022 => [("asset1", "asset2")])

weights = calculate_weight_for_investment_discounts(
    social_rate,
    discount_year,
    salvage_value,
    investment_cost,
    years,
    investable_assets,
)

# output

Dict{Tuple{Int64, Tuple{String, String}}, Float64} with 3 entries:
  (2022, ("asset1", "asset2")) => 0.0797817
  (2021, ("asset3", "asset4")) => 0.13097
  (2021, ("asset1", "asset2")) => 0.158874

TulipaEnergyModel.calculate_weight_for_investment_discounts — Method

calculate_weight_for_investment_discounts(graph::MetaGraph,
                                          years,
                                          investable_assets,
                                          assets,
                                          model_parameters,
                                         )

Calculates the weight for investment discounts for each asset, both energy assets and transport assets. Internally calls calculate_annualized_cost, calculate_salvage_value, calculate_weight_for_investment_discounts.

Arguments

graph::MetaGraph: A graph

years::Array: An array of years to be considered.

investable_assets::Dict: A dictionary where the key is a year, and the value is an array of assets that are relevant for that year.

assets::Array: An array of assets.

model_parameters::ModelParameters: A model parameters structure.

Returns

A Dict where the keys are tuples (year, asset) representing the year and the asset, and the values are the weights for investment discounts.

TulipaEnergyModel.compute_dual_variables! — Method

compute_dual_variables!(model)

Compute the dual variables for the given model.

If the model does not have dual variables, this function fixes the discrete variables, optimizes the model, and then computes the dual variables.

Arguments

model: The model for which to compute the dual variables.

Returns

A named tuple containing the dual variables of selected constraints.

TulipaEnergyModel.create_internal_structures — Method

graph, representative_periods, timeframe  = create_internal_structures(connection)

Return the graph, representative_periods, and timeframe structures given the input dataframes structure.

The details of these structures are:

graph: a MetaGraph with the following information:

labels(graph): All assets.

edge_labels(graph): All flows, in pair format (u, v), where u and v are assets.

graph[a]: A TulipaEnergyModel.GraphAssetData structure for asset a.

graph[u, v]: A TulipaEnergyModel.GraphFlowData structure for flow (u, v).

representative_periods: An array of TulipaEnergyModel.RepresentativePeriod ordered by their IDs.

timeframe: Information of TulipaEnergyModel.Timeframe.

TulipaEnergyModel.create_intervals_for_years — Method

create_intervals(years)

Create a dictionary of intervals for years. The interval is assigned to the its starting year. The last interval is 1.

TulipaEnergyModel.create_model! — Method

create_model!(energy_problem; verbose = false)

Create the internal model of an TulipaEnergyModel.EnergyProblem.

TulipaEnergyModel.create_model — Method

model = create_model(args...; write_lp_file = false, enable_names = true)

Create the energy model manually. We recommend using create_model! instead.

TulipaEnergyModel.default_parameters — Method

default_parameters(Val(optimizer_name_symbol))
default_parameters(optimizer)
default_parameters(optimizer_name_symbol)
default_parameters(optimizer_name_string)

Returns the default parameters for a given JuMP optimizer. Falls back to Dict() for undefined solvers.

Arguments

There are four ways to use this function:

Val(optimizer_name_symbol): This uses type dispatch with the special Val type. Pass the solver name as a Symbol (e.g., Val(:HiGHS)).

optimizer: The JuMP optimizer type (e.g., HiGHS.Optimizer).

optimizer_name_symbol or optimizer_name_string: Pass the name in Symbol or String format and it will be converted to Val.

Using Val is necessary for the dispatch. All other cases will convert the argument and call the Val version, which might lead to type instability.

Examples

using HiGHS
default_parameters(HiGHS.Optimizer)

# output

Dict{String, Any} with 1 entry:
  "output_flag" => false

Another case

default_parameters(Val(:Cbc))

# output

Dict{String, Any} with 1 entry:
  "logLevel" => 0

default_parameters(:Cbc) == default_parameters("Cbc") == default_parameters(Val(:Cbc))

# output

true

TulipaEnergyModel.duration — Method

Δ = duration(block, rp, representative_periods)

Computes the duration of the block and multiply by the resolution of the representative period rp.

TulipaEnergyModel.export_solution_to_csv_files — Method

export_solution_to_csv_files(output_file, connection, variables, constraints)

Saves the solution in CSV files inside output_folder. Notice that this assumes that the solution has been computed by save_solution!.

TulipaEnergyModel.export_solution_to_csv_files — Method

export_solution_to_csv_files(output_folder, energy_problem)

Saves the solution from energy_problem in CSV files inside output_file. Notice that this assumes that the solution has been computed by save_solution!.

TulipaEnergyModel.filter_graph — Method

filter_graph(graph, elements, value, key)
filter_graph(graph, elements, value, key, year)

Helper function to filter elements (assets or flows) in the graph given a key (and possibly year) and value (or values). In the safest case, this is equivalent to the filters

filter_assets_whose_key_equal_to_value = a -> graph[a].key == value
filter_assets_whose_key_year_equal_to_value = a -> graph[a].key[year] in value
filter_flows_whose_key_equal_to_value = f -> graph[f...].key == value
filter_flows_whose_key_year_equal_to_value = f -> graph[f...].key[year] in value

TulipaEnergyModel.get_graph_value_or_missing — Method

get_graph_value_or_missing(graph, graph_key, field_key)
get_graph_value_or_missing(graph, graph_key, field_key, year)

Get graph[graph_key].field_key (or graph[graph_key].field_key[year]) or return missing if any of the values do not exist. We also check if graph[graph_key].active[year] is true if the year is passed and return missing otherwise.

TulipaEnergyModel.is_active — Method

is_active(graph, a, y)
is_active(graph, (u, v), y)

Returns graph[a].active[y][y] or graph[u, v].active[y][y] or false if intermediary values are missing.

TulipaEnergyModel.profile_aggregation — Method

profile_aggregation(agg, profiles, key, block, default_value)

Aggregates the profiles[key] over the block using the agg function. If the profile does not exist, uses default_value instead of each profile value.

profiles should be a dictionary of profiles, for instance graph[a].profiles or graph[u, v].profiles. If profiles[key] exists, then this function computes the aggregation of profiles[key] over the range block using the aggregator agg, i.e., agg(profiles[key][block]). If profiles[key] does not exist, then this substitutes it with a vector of default_values.

TulipaEnergyModel.read_parameters_from_file — Method

read_parameters_from_file(filepath)

Parse the parameters from a file into a dictionary. The keys and values are NOT checked to be valid parameters for any specific solvers.

The file should contain a list of lines of the following type:

key = value

The file is parsed as TOML, which is intuitive. See the example below.

Example

# Creating file
filepath, io = mktemp()
println(io,
  """
    true_or_false = true
    integer_number = 5
    real_number1 = 3.14
    big_number = 6.66E06
    small_number = 1e-8
    string = "something"
  """
)
close(io)
# Reading
read_parameters_from_file(filepath)

# output

Dict{String, Any} with 6 entries:
  "string"         => "something"
  "integer_number" => 5
  "small_number"   => 1.0e-8
  "true_or_false"  => true
  "real_number1"   => 3.14
  "big_number"     => 6.66e6

TulipaEnergyModel.run_scenario — Method

energy_problem = run_scenario(connection; optimizer, parameters, write_lp_file, log_file, show_log)

Run the scenario in the given connection and return the energy problem.

The optimizer and parameters keyword arguments can be used to change the optimizer (the default is HiGHS) and its parameters. The variables are passed to the solve_model function.

Set write_lp_file = true to export the problem that is sent to the solver to a file for viewing. Set enable_names = false to turn off variable and constraint names (faster model creation). Set show_log = false to silence printing the log while running. Specify a log_file name to export the log to a file.

TulipaEnergyModel.safe_comparison — Method

safe_comparison(graph, a, value, key)
safe_comparison(graph, a, value, key, year)

Check if graph[a].value (or graph[a].value[year]) is equal to value. This function assumes that if graph[a].value is a dictionary and value is not, then you made a mistake. This makes it safer, because it will not silently return false. It also checks for missing.

TulipaEnergyModel.safe_inclusion — Method

safe_inclusion(graph, a, value, key)
safe_inclusion(graph, a, value, key, year)

Check if graph[a].value (or graph[a].value[year]) is in values. This correctly check that missing in [missing] returns false.

TulipaEnergyModel.save_solution! — Method

save_solution!(connection, model, variables, constraints; compute_duals = true)

Saves the primal and dual variables values, in the following way:

For each variable in variables, get the solution value and save it in a

column named solution in the corresponding dataset.

For each constraint in constraints, get the dual of each attached

constraint listed in constraint.constraint_names and save it to the dictionary constraint.duals with the key given by the name.

Notice that the duals are only computed if compute_duals is true.

TulipaEnergyModel.save_solution! — Method

save_solution!(energy_problem; compute_duals = true)

TulipaEnergyModel.solve_model — Function

solve_model(model[, optimizer; parameters])

Solve the JuMP model. The optimizer argument should be an MILP solver from the JuMP list of supported solvers. By default we use HiGHS.

The keyword argument parameters should be passed as a list of key => value pairs. These can be created manually, obtained using default_parameters, or read from a file using read_parameters_from_file.

Examples

parameters = Dict{String,Any}("presolve" => "on", "time_limit" => 60.0, "output_flag" => true)
solve_model(model, HiGHS.Optimizer; parameters = parameters)

TulipaEnergyModel.solve_model! — Function

solve_model!(energy_problem[, optimizer; parameters, save_solution = true])

Solve the internal model of an energy_problem. If save_solution, then the solution and dual variables are computed and saved by save_solution!.

TulipaEnergyModel.tmp_create_partition_tables — Method

tmp_create_partition_tables(connection)

Create the unrolled partition tables using only tables.

The table explicit_assets_rep_periods_partitions is the explicit version of assets_rep_periods_partitions, i.e., it adds the rows not defined in that table by setting the specification to 'uniform' and the partition to '1'.

The table asset_time_resolution is the unrolled version of the table above, i.e., it takes the specification and partition and expands into a series of time blocks. The columns time_block_start and time_block_end replace the specification and partition columns.

Similarly, flow tables are created as well.

TulipaEnergyModel.tmp_example_of_flow_expression_problem — Method

connection, ep = tmp_example_of_flow_expression_problem()