[Reference]

​ TODO

Contents

​

[Reference]

Contents

​## Index ​

TulipaEnergyModel._parse_rp_partition

TulipaEnergyModel.compute_rp_partition

TulipaEnergyModel.compute_rp_partitions

TulipaEnergyModel.create_graph

TulipaEnergyModel.create_parameters_and_sets_from_file

TulipaEnergyModel.read_csv_with_schema

TulipaEnergyModel.resolution_matrix

TulipaEnergyModel.save_solution_to_file

TulipaEnergyModel.solve_model

TulipaEnergyModel._parse_rp_partition — Function

_parse_rp_partition(Val(specification), time_step_string, rp_time_steps)

Parses the timestepstring according to the specification. The representative period time steps (rp_time_steps) might not be used in the computation, but it will be used for validation.

The specification defines what is expected from the time_step_string:

:uniform: The time_step_string should be a single number indicating the duration of each block. Examples: "3", "4", "1".

:explicit: The time_step_string should be a semicolon-separated list of integers. Each integer is a duration of a block. Examples: "3;3;3;3", "4;4;4", "1;1;1;1;1;1;1;1;1;1;1;1", and "3;3;4;2".

:math: The time_step_string should be an expression of the form NxD+NxD…, where D is the duration of the block and N is the number of blocks. Examples: "4x3", "3x4", "12x1", and "2x3+1x4+1x2".

The generated blocks will be ranges (a:b). The first block starts at 1, and the last block ends at length(rp_time_steps).

The following table summarizes the formats for a rp_time_steps = 1:12:

Output	:uniform	:explicit	:math

1:3, 4:6, 7:9, 10:12	3	3;3;3;3	4x3

1:4, 5:8, 9:12	4	4;4;4	3x4

1:1, 2:2, …, 12:12	1	1;1;1;1;1;1;1;1;1;1;1;1	12x1

1:3, 4:6, 7:10, 11:12	NA	3;3;4;2	2x3+1x4+1x2

Examples

using TulipaEnergyModel
TulipaEnergyModel._parse_rp_partition(Val(:uniform), "3", 1:12)

# output

4-element Vector{UnitRange{Int64}}:
 1:3
 4:6
 7:9
 10:12

using TulipaEnergyModel
TulipaEnergyModel._parse_rp_partition(Val(:explicit), "4;4;4", 1:12)

# output

3-element Vector{UnitRange{Int64}}:
 1:4
 5:8
 9:12

using TulipaEnergyModel
TulipaEnergyModel._parse_rp_partition(Val(:math), "2x3+1x4+1x2", 1:12)

# output

4-element Vector{UnitRange{Int64}}:
 1:3
 4:6
 7:10
 11:12

TulipaEnergyModel.compute_rp_partition — Method

rp_partition = compute_rp_partition(partitions; strategy = :greedy)

Given the time steps of various flows/assets in the partitions input, compute the representative period partitions.

Each element of partitions is a partition with the following assumptions:

An element is of the form V = [r₁, r₂, …, rₘ], where each rᵢ is a range a:b.

r₁ starts at 1.

rᵢ₊₁ starts at the end of rᵢ plus 1.

rₘ ends at some value N, that is the same for all elements of partitions.

Notice that this implies that they form a disjunct partition of 1:N.

The output will also be a partition with the conditions above.

Strategies

:greedy

If strategy = :greedy (default), then the output is constructed greedily, i.e., it selects the next largest breakpoint following the algorithm below:

Input: Vᴵ₁, …, Vᴵₚ, a list of time blocks. Each element of Vᴵⱼ is a range r = r.start:r.end. Output: V.

Compute the end of the representative period N (all Vᴵⱼ should have the same end)

Start with an empty V = []

Define the beginning of the range s = 1

Define an array with all the next breakpoints B such that Bⱼ is the first r.end such that r.end ≥ s for each r ∈ Vᴵⱼ.

The end of the range will be the e = max Bⱼ.

Define r = s:e and add r to the end of V.

If e = N, then END

Otherwise, define s = e + 1 and go to step 4.

Examples

partition1 = [1:4, 5:8, 9:12]
partition2 = [1:3, 4:6, 7:9, 10:12]
compute_rp_partition([partition1, partition2])

# output

3-element Vector{UnitRange{Int64}}:
 1:4
 5:8
 9:12

partition1 = [1:1, 2:3, 4:6, 7:10, 11:12]
partition2 = [1:2, 3:4, 5:5, 6:7, 8:9, 10:12]
compute_rp_partition([partition1, partition2])

# output

5-element Vector{UnitRange{Int64}}:
 1:2
 3:4
 5:6
 7:10
 11:12

:all

If strategy = :all, then the output selects includes all the breakpoints from the input. Another way of describing it, is to select the minimum end-point instead of the maximum end-point in the :greedy strategy.

Examples

partition1 = [1:4, 5:8, 9:12]
partition2 = [1:3, 4:6, 7:9, 10:12]
compute_rp_partition([partition1, partition2]; strategy = :all)

# output

6-element Vector{UnitRange{Int64}}:
 1:3
 4:4
 5:6
 7:8
 9:9
 10:12

partition1 = [1:1, 2:3, 4:6, 7:10, 11:12]
partition2 = [1:2, 3:4, 5:5, 6:7, 8:9, 10:12]
compute_rp_partition([partition1, partition2]; strategy = :all)

# output

10-element Vector{UnitRange{Int64}}:
 1:1
 2:2
 3:3
 4:4
 5:5
 6:6
 7:7
 8:9
 10:10
 11:12

TulipaEnergyModel.compute_rp_partitions — Method

compute_rp_partitions(df, elements, time_steps_per_rp)

For each element in elements (assets or flows), parse the representative period partitions defined in df, a DataFrame of the file assets-partitions.csv or flows-partitions.csv.

time_steps_per_rp must be a dictionary indexed by rp and its values are the time steps of that rp.

To obtain the partitions, the columns specification and partition from df are passed to the function _parse_rp_partition.

TulipaEnergyModel.create_graph — Method

graph = create_graph(assets_path, flows_path)

Read the assets and flows data CSVs and create a graph object.

TulipaEnergyModel.create_parameters_and_sets_from_file — Method

parameters, sets = create_parameters_and_sets_from_file(input_folder)

Returns two NamedTuples with all parameters and sets read and created from the input files in the input_folder.

TulipaEnergyModel.read_csv_with_schema — Method

read_csv_with_schema(file_path, schema)

Reads the csv with file_name at location path validating the data using the schema. It is assumes that the file's header is at the second row. The first row of the file contains some metadata information that is not used.

TulipaEnergyModel.resolution_matrix — Method

M = resolution_matrix(rp_partition, time_blocks; rp_resolution = 1.0)

Computes the resolution balance matrix using the array of rp_partition and the array of time_blocks. The time_blocks will normally be from an asset or flow, but there is nothing constraining it to that. The elements in these arrays must be ranges.

The resulting matrix will be multiplied by rp_resolution.

Examples

The following two examples are for two flows/assets with resolutions of 3h and 4h, so that the representative period has 4h periods.

rp_partition = [1:4, 5:8, 9:12]
time_blocks = [1:4, 5:8, 9:12]
resolution_matrix(rp_partition, time_blocks)

# output

3×3 SparseArrays.SparseMatrixCSC{Float64, Int64} with 3 stored entries:
 1.0   ⋅    ⋅
  ⋅   1.0   ⋅
  ⋅    ⋅   1.0

rp_partition = [1:4, 5:8, 9:12]
time_blocks = [1:3, 4:6, 7:9, 10:12]
resolution_matrix(rp_partition, time_blocks; rp_resolution = 1.5)

# output

3×4 SparseArrays.SparseMatrixCSC{Float64, Int64} with 6 stored entries:
 1.5  0.5   ⋅    ⋅
  ⋅   1.0  1.0   ⋅
  ⋅    ⋅   0.5  1.5

TulipaEnergyModel.save_solution_to_file — Method

save_solution_to_file(output_file, v_investment, unit_capacity)

Saves the solution variable vinvestment to a file "investments.csv" inside `outputfile. The format of each row isa,v,p*v, whereais the asset indexingvinvestment,vis correspondingvinvestmentvalue, andpis the correspondingunit_capacity` value.

TulipaEnergyModel.solve_model — Method

solve_model(model)

Solve the model.