[Reference]

​ TODO

Contents

​

[Reference]

Contents

​## Index ​

TulipaEnergyModel.create_graph

TulipaEnergyModel.create_parameters_and_sets_from_file

TulipaEnergyModel.get_df

TulipaEnergyModel.optimise_investments

TulipaEnergyModel.save_solution_to_file

TulipaEnergyModel.create_graph — Method

graph = create_graph(nodes_path, edges_path)

Read the nodes and edges data CSVs and create a graph object.

TulipaEnergyModel.create_parameters_and_sets_from_file — Method

parameters, sets = create_parameters_and_sets_from_file(input_folder)

Returns two NamedTuples with all parameters and sets read and created from the input files in the input_folder.

TulipaEnergyModel.get_df — Method

get_df(path, file_name, schema)

Reads the csv with file_name at location path, then validates the data using the schema.

TulipaEnergyModel.optimise_investments — Method

optimise_investments(graph, params, sets; verbose = false)

Create and solve the model using the graph structure, the parameters and sets.

TulipaEnergyModel.save_solution_to_file — Method

save_solution_to_file(output_file, v_investment, unit_capacity)

Saves the solution variable vinvestment to a file "investments.csv" inside `outputfile. The format of each row isa,v,p*v, whereais the asset indexingvinvestment,vis correspondingvinvestmentvalue, andpis the correspondingunit_capacity` value.