Reference

​This is a comprehensive list of all of the functions in the model.
The function most pertinent to the user is: run_scenario

Index

TulipaEnergyModel.AssetData

TulipaEnergyModel.AssetPartitionData

TulipaEnergyModel.AssetProfiles

TulipaEnergyModel.EnergyProblem

TulipaEnergyModel.FlowData

TulipaEnergyModel.FlowPartitionData

TulipaEnergyModel.FlowProfiles

TulipaEnergyModel.GraphAssetData

TulipaEnergyModel.GraphFlowData

TulipaEnergyModel.RepPeriodData

TulipaEnergyModel.RepresentativePeriod

TulipaEnergyModel._parse_rp_partition

TulipaEnergyModel.compute_assets_partitions!

TulipaEnergyModel.compute_constraints_partitions

TulipaEnergyModel.compute_flows_partitions!

TulipaEnergyModel.compute_rp_partition

TulipaEnergyModel.create_energy_problem_from_csv_folder

TulipaEnergyModel.create_graph_and_representative_periods_from_csv_folder

TulipaEnergyModel.create_model

TulipaEnergyModel.create_model!

TulipaEnergyModel.read_csv_with_schema

TulipaEnergyModel.resolution_matrix

TulipaEnergyModel.run_scenario

TulipaEnergyModel.save_solution_to_file

TulipaEnergyModel.save_solution_to_file

TulipaEnergyModel.solve_model

TulipaEnergyModel.solve_model!

Contents

TulipaEnergyModel.AssetData — Type

Schema for the assets-data.csv file.

TulipaEnergyModel.AssetPartitionData — Type

Schema for the assets-partitions.csv file.

TulipaEnergyModel.AssetProfiles — Type

Schema for the assets-profiles.csv file.

TulipaEnergyModel.EnergyProblem — Type

Structure to hold all parts of an energy problem.

TulipaEnergyModel.FlowData — Type

Schema for the flows-data.csv file.

TulipaEnergyModel.FlowPartitionData — Type

Schema for the flows-partitions.csv file.

TulipaEnergyModel.FlowProfiles — Type

Schema for the flows-profiles file.

TulipaEnergyModel.GraphAssetData — Type

Structure to hold the asset data in the graph.

TulipaEnergyModel.GraphFlowData — Type

Structure to hold the flow data in the graph.

TulipaEnergyModel.RepPeriodData — Type

Schema for the rep-period-data.csv file.

TulipaEnergyModel.RepresentativePeriod — Type

Structure to hold the data of one representative period.

TulipaEnergyModel._parse_rp_partition — Function

_parse_rp_partition(Val(specification), time_step_string, rp_time_steps)

Parses the timestepstring according to the specification. The representative period time steps (rp_time_steps) might not be used in the computation, but it will be used for validation.

The specification defines what is expected from the time_step_string:

:uniform: The time_step_string should be a single number indicating the duration of each block. Examples: "3", "4", "1".

:explicit: The time_step_string should be a semicolon-separated list of integers. Each integer is a duration of a block. Examples: "3;3;3;3", "4;4;4", "1;1;1;1;1;1;1;1;1;1;1;1", and "3;3;4;2".

:math: The time_step_string should be an expression of the form NxD+NxD…, where D is the duration of the block and N is the number of blocks. Examples: "4x3", "3x4", "12x1", and "2x3+1x4+1x2".

The generated blocks will be ranges (a:b). The first block starts at 1, and the last block ends at length(rp_time_steps).

The following table summarizes the formats for a rp_time_steps = 1:12:

Output	:uniform	:explicit	:math

1:3, 4:6, 7:9, 10:12	3	3;3;3;3	4x3

1:4, 5:8, 9:12	4	4;4;4	3x4

1:1, 2:2, …, 12:12	1	1;1;1;1;1;1;1;1;1;1;1;1	12x1

1:3, 4:6, 7:10, 11:12	NA	3;3;4;2	2x3+1x4+1x2

Examples

using TulipaEnergyModel
TulipaEnergyModel._parse_rp_partition(Val(:uniform), "3", 1:12)

# output

4-element Vector{UnitRange{Int64}}:
 1:3
 4:6
 7:9
 10:12

using TulipaEnergyModel
TulipaEnergyModel._parse_rp_partition(Val(:explicit), "4;4;4", 1:12)

# output

3-element Vector{UnitRange{Int64}}:
 1:4
 5:8
 9:12

using TulipaEnergyModel
TulipaEnergyModel._parse_rp_partition(Val(:math), "2x3+1x4+1x2", 1:12)

# output

4-element Vector{UnitRange{Int64}}:
 1:3
 4:6
 7:10
 11:12

TulipaEnergyModel.compute_assets_partitions! — Method

compute_assets_partitions!(partitions, df, a, representative_periods)

Parses the time blocks in the DataFrame df for the asset a and every representative period in the time_steps_per_rp dictionary, modifying the input partitions.

partitions must be a dictionary indexed by the representative periods, possibly empty.

time_steps_per_rp must be a dictionary indexed by rp and its values are the time steps of that rp.

To obtain the partitions, the columns specification and partition from df are passed to the function _parse_rp_partition.

TulipaEnergyModel.compute_constraints_partitions — Method

cons_partitions = compute_constraints_partitions(graph, representative_periods)

Computes the constraints partitions using the assets and flows partitions stored in the graph, and the representative periods.

The function computes the constraints partitions by iterating over the partition dictionary, which specifies the partition strategy for each resolution (i.e., lowest or highest). For each asset and representative period, it calls the compute_rp_partition function to compute the partition based on the strategy.

TulipaEnergyModel.compute_flows_partitions! — Method

compute_flows_partitions!(partitions, df, u, v, representative_periods)

Parses the time blocks in the DataFrame df for the flow (u, v) and every representative period in the time_steps_per_rp dictionary, modifying the input partitions.

partitions must be a dictionary indexed by the representative periods, possibly empty.

time_steps_per_rp must be a dictionary indexed by rp and its values are the time steps of that rp.

To obtain the partitions, the columns specification and partition from df are passed to the function _parse_rp_partition.

TulipaEnergyModel.compute_rp_partition — Method

rp_partition = compute_rp_partition(partitions; strategy = :greedy)

Given the time steps of various flows/assets in the partitions input, compute the representative period partitions.

Each element of partitions is a partition with the following assumptions:

An element is of the form V = [r₁, r₂, …, rₘ], where each rᵢ is a range a:b.

r₁ starts at 1.

rᵢ₊₁ starts at the end of rᵢ plus 1.

rₘ ends at some value N, that is the same for all elements of partitions.

Notice that this implies that they form a disjunct partition of 1:N.

The output will also be a partition with the conditions above.

Strategies

:greedy

If strategy = :greedy (default), then the output is constructed greedily, i.e., it selects the next largest breakpoint following the algorithm below:

Input: Vᴵ₁, …, Vᴵₚ, a list of time blocks. Each element of Vᴵⱼ is a range r = r.start:r.end. Output: V.

Compute the end of the representative period N (all Vᴵⱼ should have the same end)

Start with an empty V = []

Define the beginning of the range s = 1

Define an array with all the next breakpoints B such that Bⱼ is the first r.end such that r.end ≥ s for each r ∈ Vᴵⱼ.

The end of the range will be the e = max Bⱼ.

Define r = s:e and add r to the end of V.

If e = N, then END

Otherwise, define s = e + 1 and go to step 4.

Examples

partition1 = [1:4, 5:8, 9:12]
partition2 = [1:3, 4:6, 7:9, 10:12]
compute_rp_partition([partition1, partition2])

# output

3-element Vector{UnitRange{Int64}}:
 1:4
 5:8
 9:12

partition1 = [1:1, 2:3, 4:6, 7:10, 11:12]
partition2 = [1:2, 3:4, 5:5, 6:7, 8:9, 10:12]
compute_rp_partition([partition1, partition2])

# output

5-element Vector{UnitRange{Int64}}:
 1:2
 3:4
 5:6
 7:10
 11:12

:all

If strategy = :all, then the output selects includes all the breakpoints from the input. Another way of describing it, is to select the minimum end-point instead of the maximum end-point in the :greedy strategy.

Examples

partition1 = [1:4, 5:8, 9:12]
partition2 = [1:3, 4:6, 7:9, 10:12]
compute_rp_partition([partition1, partition2]; strategy = :all)

# output

6-element Vector{UnitRange{Int64}}:
 1:3
 4:4
 5:6
 7:8
 9:9
 10:12

partition1 = [1:1, 2:3, 4:6, 7:10, 11:12]
partition2 = [1:2, 3:4, 5:5, 6:7, 8:9, 10:12]
compute_rp_partition([partition1, partition2]; strategy = :all)

# output

10-element Vector{UnitRange{Int64}}:
 1:1
 2:2
 3:3
 4:4
 5:5
 6:6
 7:7
 8:9
 10:10
 11:12

TulipaEnergyModel.create_energy_problem_from_csv_folder — Method

energy_problem = create_energy_problem_from_csv_folder(input_folder)

Returns the TulipaEnergyModel.EnergyProblem reading all data from CSV files in the input_folder. This is a wrapper around create_graph_and_representative_periods_from_csv_folder that creates the EnergyProblem structure.

TulipaEnergyModel.create_graph_and_representative_periods_from_csv_folder — Method

graph, representative_periods = create_graph_and_representative_periods_from_csv_folder(input_folder)

Returns the graph structure that holds all data, and the representative_periods array. The following files are expected to exist in the input folder:

assets-data.csv: Following the TulipaEnergyModel.AssetData specification.

assets-profiles.csv: Following the TulipaEnergyModel.AssetProfiles specification. The profiles should be ordered by time step.

assets-paritions.csv: Following the TulipaEnergyModel.AssetPartitionData specification.

flows-data.csv: Following the TulipaEnergyModel.FlowData specification.

flows-profiles.csv: Following the TulipaEnergyModel.FlowProfiles specification. The profiles should be ordered by time step.

flows-paritions.csv: Following the TulipaEnergyModel.FlowPartitionData specification.

rep-periods-data.csv: Following the TulipaEnergyModel.RepPeriodData specification.

The returned structures are:

graph: a MetaGraph with the following information:

labels(graph): All assets.

edge_labels(graph): All flows, in pair format (u, v), where u and v are assets.

graph[a]: A TulipaEnergyModel.GraphAssetData structure for asset a.

graph[u, v]: A TulipaEnergyModel.GraphFlowData structure for flow (u, v).

representative_periods: An array of TulipaEnergyModel.RepresentativePeriod ordered by their IDs.

TulipaEnergyModel.create_model! — Method

create_model!(energy_problem; verbose = false)

Create the internal model of an TulipaEnergyModel.EnergyProblem.

TulipaEnergyModel.create_model — Method

model = create_model(graph, representative_periods)

Create the energy model given the graph and representative_periods.

TulipaEnergyModel.read_csv_with_schema — Method

read_csv_with_schema(file_path, schema)

Reads the csv with file_name at location path validating the data using the schema. It assumes that the file's header is at the second row. The first row of the file contains some metadata information that is not used.

TulipaEnergyModel.resolution_matrix — Method

M = resolution_matrix(rp_partition, time_blocks; rp_resolution = 1.0)

Computes the resolution balance matrix using the array of rp_partition and the array of time_blocks. The time_blocks will normally be from an asset or flow, but there is nothing constraining it to that. The elements in these arrays must be ranges.

The resulting matrix will be multiplied by rp_resolution.

Examples

The following two examples are for two flows/assets with resolutions of 3h and 4h, so that the representative period has 4h periods.

rp_partition = [1:4, 5:8, 9:12]
time_blocks = [1:4, 5:8, 9:12]
resolution_matrix(rp_partition, time_blocks)

# output

3×3 SparseArrays.SparseMatrixCSC{Float64, Int64} with 3 stored entries:
 1.0   ⋅    ⋅
  ⋅   1.0   ⋅
  ⋅    ⋅   1.0

rp_partition = [1:4, 5:8, 9:12]
time_blocks = [1:3, 4:6, 7:9, 10:12]
resolution_matrix(rp_partition, time_blocks; rp_resolution = 1.5)

# output

3×4 SparseArrays.SparseMatrixCSC{Float64, Int64} with 6 stored entries:
 1.5  0.5   ⋅    ⋅
  ⋅   1.0  1.0   ⋅
  ⋅    ⋅   0.5  1.5

TulipaEnergyModel.run_scenario — Method

run_scenario(input_folder)
run_scenario(input_folder, output_folder)

Run the scenario in the given inputfolder and return the sets, parameters, and solution. If the outputfolder is specified, save the sets, parameters, and solution to the output_folder.

TulipaEnergyModel.save_solution_to_file — Method

save_solution_to_file(output_file, graph)

Saves the solution in CSV files inside output_folder.

The following files are created:

investment.csv: The format of each row is a,v,p*v, where a is the asset name, v is the corresponding asset investment value, and p is the corresponding capacity value. Only investable assets are included.

TulipaEnergyModel.save_solution_to_file — Method

save_solution_to_file(output_folder, energy_problem)

Saves the solution from energy_problem in CSV files inside output_file.

TulipaEnergyModel.solve_model! — Method

solution = solve_model!(energy_problem)

Solve the internal model of an energyproblem. The solution obtained by calling [`solvemodel`](@ref) is returned.

TulipaEnergyModel.solve_model — Method

solution = solve_model(model)

Solve the JuMP model and return the solution.

The solution object is a NamedTuple with the following fields:

objective_value: A Float64 with the objective value at the solution.

assets_investment[a]: The investment for each asset, indexed on the investable asset a. To create a traditional array in the order given by the investable assets, one can run

[solution.assets_investment[a] for a in labels(graph) if graph[a].investable]

flows_investment[u, v]: The investment for each flow, indexed on the investable flow (u, v). To create a traditional array in the order given by the investable flows, one can run

[solution.flows_investment[(u, v)] for (u, v) in edge_labels(graph) if graph[u, v].investable]

flow[(u, v), rp, B]: The flow value for a given flow (u, v) at a given representative period rp, and time block B. The list of time blocks is defined by graph[(u, v)].partitions[rp]. To create a vector with all values of flow for a given (u, v) and rp, one can run

[solution.flow[(u, v), rp, B] for B in graph[u, v].partitions[rp]]

storage_level[a, rp, B]: The storage level for the storage asset a for a representative period rp and a time block B. The list of time blocks is defined by constraints_partitions, which was used to create the model. To create a vector with the all values of storage_level for a given a and rp, one can run

[solution.storage_level[a, rp, B] for B in constraints_partitions[:lowest_resolution][(a, rp)]]