Reference

​This is a comprehensive list of all of the functions in the model.
The function most pertinent to the user is: run_scenario

Index

TulipaEnergyModel.AssetData

TulipaEnergyModel.AssetPartitionData

TulipaEnergyModel.AssetProfiles

TulipaEnergyModel.BasePeriod

TulipaEnergyModel.EnergyProblem

TulipaEnergyModel.FlowData

TulipaEnergyModel.FlowPartitionData

TulipaEnergyModel.FlowProfiles

TulipaEnergyModel.GraphAssetData

TulipaEnergyModel.GraphFlowData

TulipaEnergyModel.RepPeriodData

TulipaEnergyModel.RepPeriodMapping

TulipaEnergyModel.RepresentativePeriod

TulipaEnergyModel._parse_rp_partition

TulipaEnergyModel.add_expression_terms!

TulipaEnergyModel.compute_assets_partitions!

TulipaEnergyModel.compute_constraints_partitions

TulipaEnergyModel.compute_dual_variables

TulipaEnergyModel.compute_flows_partitions!

TulipaEnergyModel.compute_rp_partition

TulipaEnergyModel.construct_dataframes

TulipaEnergyModel.create_energy_problem_from_csv_folder

TulipaEnergyModel.create_graph_and_representative_periods_from_csv_folder

TulipaEnergyModel.create_model

TulipaEnergyModel.create_model!

TulipaEnergyModel.default_parameters

TulipaEnergyModel.profile_aggregation

TulipaEnergyModel.read_csv_with_schema

TulipaEnergyModel.read_parameters_from_file

TulipaEnergyModel.run_scenario

TulipaEnergyModel.save_solution_to_file

TulipaEnergyModel.save_solution_to_file

TulipaEnergyModel.solve_model

TulipaEnergyModel.solve_model!

TulipaEnergyModel.solve_model!

Contents

TulipaEnergyModel.AssetData — Type

Schema for the assets-data.csv file.

TulipaEnergyModel.AssetPartitionData — Type

Schema for the assets-partitions.csv file.

TulipaEnergyModel.AssetProfiles — Type

Schema for the assets-profiles.csv file.

TulipaEnergyModel.BasePeriod — Type

Structure to hold the data of the base periods.

TulipaEnergyModel.EnergyProblem — Type

Structure to hold all parts of an energy problem. It is a wrapper around various other relevant structures. It hides the complexity behind the energy problem, making the usage more friendly, although more verbose.

Fields

graph: The Graph object that defines the geometry of the energy problem.

representative_periods: A vector of Representative Periods.

constraints_partitions: Dictionaries that connect pairs of asset and representative periods to time partitions (vectors of time blocks)

base_periods: The number of periods of the representative_periods.

dataframes: The data frames used to linearize the variables and constraints. These are used internally in the model only.

model: A JuMP.Model object representing the optimization model.

solved: A boolean indicating whether the model has been solved or not.

objective_value: The objective value of the solved problem.

termination_status: The termination status of the optimization model.

time_read_data: Time taken for reading the data (in seconds).

time_create_model: Time taken for creating the model (in seconds).

time_solve_model: Time taken for solving the model (in seconds).

Constructor

EnergyProblem(graph, representative_periods, base_periods): Constructs a new EnergyProblem object with the given graph, representative periods, and base periods. The constraints_partitions field is computed from the representative_periods, and the other fields are initialized with default values.

See the basic example tutorial to see how these can be used.

TulipaEnergyModel.FlowData — Type

Schema for the flows-data.csv file.

TulipaEnergyModel.FlowPartitionData — Type

Schema for the flows-partitions.csv file.

TulipaEnergyModel.FlowProfiles — Type

Schema for the flows-profiles file.

TulipaEnergyModel.GraphAssetData — Type

Structure to hold the asset data in the graph.

TulipaEnergyModel.GraphFlowData — Type

Structure to hold the flow data in the graph.

TulipaEnergyModel.RepPeriodData — Type

Schema for the rep-period-data.csv file.

TulipaEnergyModel.RepPeriodMapping — Type

Schema for the rep-periods-mapping.csv file.

TulipaEnergyModel.RepresentativePeriod — Type

Structure to hold the data of one representative period.

TulipaEnergyModel._parse_rp_partition — Function

_parse_rp_partition(Val(specification), time_step_string, rp_time_steps)

Parses the timestepstring according to the specification. The representative period time steps (rp_time_steps) might not be used in the computation, but it will be used for validation.

The specification defines what is expected from the time_step_string:

:uniform: The time_step_string should be a single number indicating the duration of each block. Examples: "3", "4", "1".

:explicit: The time_step_string should be a semicolon-separated list of integers. Each integer is a duration of a block. Examples: "3;3;3;3", "4;4;4", "1;1;1;1;1;1;1;1;1;1;1;1", and "3;3;4;2".

:math: The time_step_string should be an expression of the form NxD+NxD…, where D is the duration of the block and N is the number of blocks. Examples: "4x3", "3x4", "12x1", and "2x3+1x4+1x2".

The generated blocks will be ranges (a:b). The first block starts at 1, and the last block ends at length(rp_time_steps).

The following table summarizes the formats for a rp_time_steps = 1:12:

Output	:uniform	:explicit	:math

1:3, 4:6, 7:9, 10:12	3	3;3;3;3	4x3

1:4, 5:8, 9:12	4	4;4;4	3x4

1:1, 2:2, …, 12:12	1	1;1;1;1;1;1;1;1;1;1;1;1	12x1

1:3, 4:6, 7:10, 11:12	NA	3;3;4;2	2x3+1x4+1x2

Examples

using TulipaEnergyModel
TulipaEnergyModel._parse_rp_partition(Val(:uniform), "3", 1:12)

# output

4-element Vector{UnitRange{Int64}}:
 1:3
 4:6
 7:9
 10:12

using TulipaEnergyModel
TulipaEnergyModel._parse_rp_partition(Val(:explicit), "4;4;4", 1:12)

# output

3-element Vector{UnitRange{Int64}}:
 1:4
 5:8
 9:12

using TulipaEnergyModel
TulipaEnergyModel._parse_rp_partition(Val(:math), "2x3+1x4+1x2", 1:12)

# output

4-element Vector{UnitRange{Int64}}:
 1:3
 4:6
 7:10
 11:12

TulipaEnergyModel.add_expression_terms! — Method

add_expression_terms!(df_cons, df_flows, workspace, representative_periods; add_efficiency)

Computes the incoming and outgoing expressions per row of df_cons. This function is only used internally in the model.

This strategy is based on the replies in this discourse thread:

https://discourse.julialang.org/t/help-improving-the-speed-of-a-dataframes-operation/107615/23

TulipaEnergyModel.compute_assets_partitions! — Method

compute_assets_partitions!(partitions, df, a, representative_periods)

Parses the time blocks in the DataFrame df for the asset a and every representative period in the time_steps_per_rp dictionary, modifying the input partitions.

partitions must be a dictionary indexed by the representative periods, possibly empty.

time_steps_per_rp must be a dictionary indexed by rp and its values are the time steps of that rp.

To obtain the partitions, the columns specification and partition from df are passed to the function _parse_rp_partition.

TulipaEnergyModel.compute_constraints_partitions — Method

cons_partitions = compute_constraints_partitions(graph, representative_periods)

Computes the constraints partitions using the assets and flows partitions stored in the graph, and the representative periods.

The function computes the constraints partitions by iterating over the partition dictionary, which specifies the partition strategy for each resolution (i.e., lowest or highest). For each asset and representative period, it calls the compute_rp_partition function to compute the partition based on the strategy.

TulipaEnergyModel.compute_dual_variables — Method

compute_dual_variables(model)

Compute the dual variables for the given model.

If the model does not have dual variables, this function fixes the discrete variables, optimizes the model, and then computes the dual variables.

Arguments

model: The model for which to compute the dual variables.

Returns

A named tuple containing the dual variables of selected constraints.

TulipaEnergyModel.compute_flows_partitions! — Method

compute_flows_partitions!(partitions, df, u, v, representative_periods)

Parses the time blocks in the DataFrame df for the flow (u, v) and every representative period in the time_steps_per_rp dictionary, modifying the input partitions.

partitions must be a dictionary indexed by the representative periods, possibly empty.

time_steps_per_rp must be a dictionary indexed by rp and its values are the time steps of that rp.

To obtain the partitions, the columns specification and partition from df are passed to the function _parse_rp_partition.

TulipaEnergyModel.compute_rp_partition — Method

rp_partition = compute_rp_partition(partitions, :lowest)

Given the time steps of various flows/assets in the partitions input, compute the representative period partitions.

Each element of partitions is a partition with the following assumptions:

An element is of the form V = [r₁, r₂, …, rₘ], where each rᵢ is a range a:b.

r₁ starts at 1.

rᵢ₊₁ starts at the end of rᵢ plus 1.

rₘ ends at some value N, that is the same for all elements of partitions.

Notice that this implies that they form a disjunct partition of 1:N.

The output will also be a partition with the conditions above.

Strategies

:lowest

If strategy = :lowest (default), then the output is constructed greedily, i.e., it selects the next largest breakpoint following the algorithm below:

Input: Vᴵ₁, …, Vᴵₚ, a list of time blocks. Each element of Vᴵⱼ is a range r = r.start:r.end. Output: V.

Compute the end of the representative period N (all Vᴵⱼ should have the same end)

Start with an empty V = []

Define the beginning of the range s = 1

Define an array with all the next breakpoints B such that Bⱼ is the first r.end such that r.end ≥ s for each r ∈ Vᴵⱼ.

The end of the range will be the e = max Bⱼ.

Define r = s:e and add r to the end of V.

If e = N, then END

Otherwise, define s = e + 1 and go to step 4.

Examples

partition1 = [1:4, 5:8, 9:12]
partition2 = [1:3, 4:6, 7:9, 10:12]
compute_rp_partition([partition1, partition2], :lowest)

# output

3-element Vector{UnitRange{Int64}}:
 1:4
 5:8
 9:12

partition1 = [1:1, 2:3, 4:6, 7:10, 11:12]
partition2 = [1:2, 3:4, 5:5, 6:7, 8:9, 10:12]
compute_rp_partition([partition1, partition2], :lowest)

# output

5-element Vector{UnitRange{Int64}}:
 1:2
 3:4
 5:6
 7:10
 11:12

:highest

If strategy = :highest, then the output selects includes all the breakpoints from the input. Another way of describing it, is to select the minimum end-point instead of the maximum end-point in the :lowest strategy.

Examples

partition1 = [1:4, 5:8, 9:12]
partition2 = [1:3, 4:6, 7:9, 10:12]
compute_rp_partition([partition1, partition2], :highest)

# output

6-element Vector{UnitRange{Int64}}:
 1:3
 4:4
 5:6
 7:8
 9:9
 10:12

partition1 = [1:1, 2:3, 4:6, 7:10, 11:12]
partition2 = [1:2, 3:4, 5:5, 6:7, 8:9, 10:12]
compute_rp_partition([partition1, partition2], :highest)

# output

10-element Vector{UnitRange{Int64}}:
 1:1
 2:2
 3:3
 4:4
 5:5
 6:6
 7:7
 8:9
 10:10
 11:12

TulipaEnergyModel.construct_dataframes — Method

dataframes = construct_dataframes(
    graph,
    representative_periods,
    constraints_partitions,
)

Computes the data frames used to linearize the variables and constraints. These are used internally in the model only.

TulipaEnergyModel.create_energy_problem_from_csv_folder — Method

energy_problem = create_energy_problem_from_csv_folder(input_folder; strict = false)

Returns the TulipaEnergyModel.EnergyProblem reading all data from CSV files in the input_folder. This is a wrapper around create_graph_and_representative_periods_from_csv_folder that creates the EnergyProblem structure. Set strict = true to error if assets are missing from partition data.

TulipaEnergyModel.create_graph_and_representative_periods_from_csv_folder — Method

graph, representative_periods, base_periods = create_graph_and_representative_periods_from_csv_folder(input_folder; strict = false)

Returns the graph structure that holds all data, and the representative_periods array. Set strict = true to error if assets are missing from partition data.

The following files are expected to exist in the input folder:

assets-data.csv: Following the TulipaEnergyModel.AssetData specification.

assets-profiles.csv: Following the TulipaEnergyModel.AssetProfiles specification. The profiles should be ordered by time step.

assets-paritions.csv: Following the TulipaEnergyModel.AssetPartitionData specification.

flows-data.csv: Following the TulipaEnergyModel.FlowData specification.

flows-profiles.csv: Following the TulipaEnergyModel.FlowProfiles specification. The profiles should be ordered by time step.

flows-paritions.csv: Following the TulipaEnergyModel.FlowPartitionData specification.

rep-periods-data.csv: Following the TulipaEnergyModel.RepPeriodData specification.

rep-periods-mapping.csv: Following the TulipaEnergyModel.RepPeriodMapping specification.

The returned structures are:

graph: a MetaGraph with the following information:

labels(graph): All assets.

edge_labels(graph): All flows, in pair format (u, v), where u and v are assets.

graph[a]: A TulipaEnergyModel.GraphAssetData structure for asset a.

graph[u, v]: A TulipaEnergyModel.GraphFlowData structure for flow (u, v).

representative_periods: An array of TulipaEnergyModel.RepresentativePeriod ordered by their IDs.

base_periods: Information of TulipaEnergyModel.BasePeriod.

TulipaEnergyModel.create_model! — Method

create_model!(energy_problem; verbose = false)

Create the internal model of an TulipaEnergyModel.EnergyProblem.

TulipaEnergyModel.create_model — Method

model = create_model(graph, representative_periods, dataframes, base_periods; write_lp_file = false)

Create the energy model given the graph, representative_periods, dictionary of dataframes (created by construct_dataframes), and base_periods.

TulipaEnergyModel.default_parameters — Method

default_parameters(Val(optimizer_name_symbol))
default_parameters(optimizer)
default_parameters(optimizer_name_symbol)
default_parameters(optimizer_name_string)

Returns the default parameters for a given JuMP optimizer. Falls back to Dict() for undefined solvers.

Arguments

There are four ways to use this function:

Val(optimizer_name_symbol): This uses type dispatch with the special Val type. Just give the solver name as a Symbol (e.g., Val(:HiGHS)).

optimizer: The JuMP optimizer type (e.g., HiGHS.Optimizer).

optimizer_name_symbol or optimizer_name_string: Just give the name in Symbol or String format and we will convert to Val.

Using Val is necessary for the dispatch. All other cases will convert the argument and call the Val version, which might lead to some type instability.

Examples

using HiGHS
default_parameters(HiGHS.Optimizer)

# output

Dict{String, Any} with 1 entry:
  "output_flag" => false

Another case

default_parameters(Val(:Cbc))

# output

Dict{String, Any} with 1 entry:
  "logLevel" => 0

default_parameters(:Cbc) == default_parameters("Cbc") == default_parameters(Val(:Cbc))

# output

true

TulipaEnergyModel.profile_aggregation — Method

profile_aggregation(agg, profiles, rp, time_block, default_value)

Aggregates the profiles[rp] over the time_block using the agg function. If the profile does not exist, uses default_value instead of each profile value.

profiles should be a dictionary of profiles, for instance graph[a].profiles or graph[u, v].profiles. If profiles[rp] exists, then this function computes the aggregation of profiles[rp] over the range time_block using the aggregator agg, i.e., agg(profiles[rp][time_block]). If profiles[rp] does not exist, then this substitutes it by a vector of default_values.

TulipaEnergyModel.read_csv_with_schema — Method

read_csv_with_schema(file_path, schema)

Reads the csv with file_name at location path validating the data using the schema. It assumes that the file's header is at the second row. The first row of the file contains some metadata information that is not used.

TulipaEnergyModel.read_parameters_from_file — Method

read_parameters_from_file(filepath)

Parse the parameters from a file into a dictionary. The keys and values are NOT checked to be valid parameters for any specific solvers.

The file should contain a list of lines of the following type:

key = value

The file is parsed as TOML, which is very intuitive. See the example below.

Example

# Creating file
filepath, io = mktemp()
println(io,
  """
    true_or_false = true
    integer_number = 5
    real_number1 = 3.14
    big_number = 6.66E06
    small_number = 1e-8
    string = "something"
  """
)
close(io)
# Reading
read_parameters_from_file(filepath)

# output

Dict{String, Any} with 6 entries:
  "string"         => "something"
  "integer_number" => 5
  "small_number"   => 1.0e-8
  "true_or_false"  => true
  "real_number1"   => 3.14
  "big_number"     => 6.66e6

TulipaEnergyModel.run_scenario — Function

energy_problem = run_scenario(input_folder[, output_folder; optimizer, parameters])

Run the scenario in the given input_folder and return the energy problem. The output_folder is optional. If it is specified, save the sets, parameters, and solution to the output_folder.

The optimizer and parameters keyword arguments can be used to change the default optimizer (which is HiGHS) and its parameters. The variables are passed to the solve_model function.

TulipaEnergyModel.save_solution_to_file — Method

save_solution_to_file(output_file, graph)

Saves the solution in CSV files inside output_folder.

The following files are created:

investment.csv: The format of each row is a,v,p*v, where a is the asset name, v is the corresponding asset investment value, and p is the corresponding capacity value. Only investable assets are included.

TulipaEnergyModel.save_solution_to_file — Method

save_solution_to_file(output_folder, energy_problem)

Saves the solution from energy_problem in CSV files inside output_file.

TulipaEnergyModel.solve_model — Function

solution = solve_model(model[, optimizer; parameters])

Solve the JuMP model and return the solution. The optimizer argument should be a MILP solver from the JuMP list of supported solvers. By default we use HiGHS.

The keyword argument parameters should be passed as a list of key => value pairs. These can be created manually, obtained using default_parameters, or read from a file using read_parameters_from_file.

The solution object is a NamedTuple with the following fields:

objective_value: A Float64 with the objective value at the solution.

assets_investment[a]: The investment for each asset, indexed on the investable asset a. To create a traditional array in the order given by the investable assets, one can run

[solution.assets_investment[a] for a in labels(graph) if graph[a].investable]

flows_investment[u, v]: The investment for each flow, indexed on the investable flow (u, v). To create a traditional array in the order given by the investable flows, one can run

[solution.flows_investment[(u, v)] for (u, v) in edge_labels(graph) if graph[u, v].investable]

flow[(u, v), rp, B]: The flow value for a given flow (u, v) at a given representative period rp, and time block B. The list of time blocks is defined by graph[(u, v)].partitions[rp]. To create a vector with all values of flow for a given (u, v) and rp, one can run

[solution.flow[(u, v), rp, B] for B in graph[u, v].partitions[rp]]

storage_level[a, rp, B]: The storage level for the storage asset a for a representative period rp and a time block B. The list of time blocks is defined by constraints_partitions, which was used to create the model. To create a vector with the all values of storage_level for a given a and rp, one can run

[solution.storage_level[a, rp, B] for B in constraints_partitions[:lowest_resolution][(a, rp)]]

Examples

parameters = Dict{String,Any}("presolve" => "on", "time_limit" => 60.0, "output_flag" => true)
solution = solve_model(model, HiGHS.Optimizer; parameters = parameters)

TulipaEnergyModel.solve_model! — Function

solution = solve_model!(energy_problem[, optimizer; parameters])

Solve the internal model of an energy_problem. The solution obtained by calling solve_model is returned.

TulipaEnergyModel.solve_model! — Method

solution = solve_model!(dataframes, model, ...)

Solves the JuMP model and return the solution, and modifies some dataframes to include the solution. The modifications made to dataframes are:

df_flows.solution = solution.flow

df_storage_level.solution = solution.storage_level